+++ src/main/java/com/kusitms/jipbap/store/dto/StoreDto.java
@@ -16,6 +16,6 @@ public class StoreDto {
     private String description;
     private Boolean koreanYn;
     private Double avgRate; // 가게 평점
-    private Long minOrderAmount; //최소 주문 금액
+    private Double minOrderAmount; //최소 주문 금액
     private String[] images;
 }
